Move engines in library
